fix casting issue

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/account/service/AccountAssociationsReadPlatformServiceImpl.java
@@ -99,14 +99,14 @@ public class AccountAssociationsReadPlatformServiceImpl implements AccountAssoci
 
         final List<Map<String, Object>> statusList = this.jdbcTemplate.queryForList(sql1, savingsId);
         for (final Map<String, Object> statusMap : statusList) {
-            AccountAssociationType associationType = AccountAssociationType.fromInt((Integer) statusMap.get("type"));
+            AccountAssociationType associationType = AccountAssociationType.fromInt(((Short) statusMap.get("type")).intValue());
             if (!associationType.isLinkedAccountAssociation() && (Boolean) statusMap.get("active")) {
                 hasActiveAccount = true;
                 break;
             }
 
             if (statusMap.get("loanStatus") != null) {
-                final LoanStatus loanStatus = LoanStatus.fromInt((Integer) statusMap.get("loanStatus"));
+                final LoanStatus loanStatus = LoanStatus.fromInt(((Short) statusMap.get("loanStatus")).intValue());
                 if (loanStatus.isActiveOrAwaitingApprovalOrDisbursal() || loanStatus.isUnderTransfer()) {
                     hasActiveAccount = true;
                     break;
@@ -114,7 +114,7 @@ public class AccountAssociationsReadPlatformServiceImpl implements AccountAssoci
             }
 
             if (statusMap.get("savingsStatus") != null) {
-                final SavingsAccountStatusType saveStatus = SavingsAccountStatusType.fromInt((Integer) statusMap.get("savingsStatus"));
+                final SavingsAccountStatusType saveStatus = SavingsAccountStatusType.fromInt(((Short) statusMap.get("savingsStatus")).intValue());
                 if (saveStatus.isActiveOrAwaitingApprovalOrDisbursal() || saveStatus.isUnderTransfer()) {
                     hasActiveAccount = true;
                     break;
